[feat] 게시글 검색 창 API 이전 및 필요한 API 명세
