+++ src/main/java/trend/project/service/planService/PlanMainPageService.java
@@ -16,6 +16,4 @@ public interface PlanMainPageService {
     List<PlanMainPageDTO.PlanCategoryResponseDTO> getTopPlanByCategory();
 
     List<PlanMainPageDTO.PlanFavoriteMemberDTO> getPlanByPopularUsers();
-
-    List<PlanMainPageDTO.PlanSearchResponseDTO> searchPlan(String title);
 }
