+++ src/main/java/trend/project/service/planService/PlanSearchService.java
@@ -1,0 +1,11 @@
+package trend.project.service.planService;
+
+import org.springframework.stereotype.Service;
+import trend.project.web.dto.planDTO.PlanMainPageDTO;
+
+import java.util.List;
+
+@Service
+public interface PlanSearchService {
+    List<PlanMainPageDTO.PlanSearchResponseDTO> searchPlan(String title);
+}
